comments_app_backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
envparse==0.2.0
filelock==3.0.12
pipenv==2020.8.13
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
